Migrate some code into new fitting_helpers sub-package
