+++ diffmah/__init__.py
@@ -6,4 +6,5 @@
 from ._version import __version__
 from .defaults import DEFAULT_MAH_PARAMS, MAH_K, DiffmahParams
 from .diffmah_kernels import mah_halopop, mah_singlehalo
-from .fitting_helpers import diffmah_fitter
+from .fitting_helpers import *
+from .fitting_helpers import bfgs_wrapper
