+++ diffmah/diffmahpop_kernels/bimod_logm0_sats/logm0_c1_late.py
@@ -7,7 +7,7 @@ from jax import jit as jjit
 from jax import numpy as jnp
 from jax import value_and_grad, vmap
 
-from ...bfgs_wrapper import bfgs_adam_fallback
+from ...fitting_helpers.bfgs_wrapper import bfgs_adam_fallback
 from ...utils import _inverse_sigmoid, _sig_slope, _sigmoid
 
 DEFAULT_LGM0POP_C1_PDICT = OrderedDict(
