+++ diffmah/fitting_helpers/__init__.py
@@ -1,0 +1,8 @@
+"""
+"""
+
+# flake8: noqa
+
+from .diffmah_fitter_helpers import diffmah_fitter
+
+__all__ = ("diffmah_fitter",)
